% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEpiValues.R
\name{getMisc}
\alias{getMisc}
\alias{getSpecies}
\title{Get extracted keywords}
\usage{
getMisc(res, full = TRUE, type = "disease")
}
\arguments{
\item{res}{the nested list resulting from \code{doc2keywords}}

\item{full}{logical, whether the \code{res} is the full results, or only one section}

\item{type}{the type of result}
}
\value{
a data.frame with
\itemize{
  \item{label: the term as it occurs in the text}
  \item{keyword/species/etc.: the canonical term matched in the dictionary}
  \item{start: the beginning of the term's span.}
  \item{end: the end of the term's span}
}
}
\description{
Extract information from extracted results
}
\details{
These functions extract information from the results of \code{doc2keywords} into
a data.frame.
}
\author{
Duncan Temple Lang
}
