\name{xml2eco}
\alias{xml2eco}
\title{XML to EcoJSON}
\description{
Takes an XML file from a PDF created using pdftohtml, and runs the
augmented keyword extractor on it. The intermediate text, broken into
sections, and the results can be cached to disk.
}
\usage{
xml2eco(XML, ecoextractLoc = getEcoExtractPyScript(),
        results_dir = character(), cache.dir = character())
}
\arguments{
  \item{XML}{the name of the XML document generated by converting the
    PDF document to XML via pdftohtml or indirectly via the
    \code{\link{convertPDF2XML}} function in ReadPDF.
}
  \item{ecoextractLoc}{the file path to the ecoextract.py
	script. Defaults to the included script in python/ecoextract.py.
}
  \item{results_dir}{
	optional directory to save the results into. Will be created if it
	does not exist.
  }
  \item{cache.dir}{
	optional directory to save the intermediate text, broken into
	sections.
}
}
\details{
  }
\value{
  A list representation of the JSON object returned by ecoextract.py
  \item{location}{}
  \item{date}{}
  \item{resolved_keyword}{}
  \item{txt}{original text}
}


\references{
}
\author{
Matt Espe
}
\note{
}


\seealso{
}
\examples{
f = system.file("sampleDocs", "Rondini-2008-Development of multiplex PCR-liga.xml", package = "SpilloverDA")
tmp = xml2eco(f)
}
\keyword{}
